type-c-service/tps6699x: Add workaround for autofet sink rejection
When the controller is running FW with autofet enabled, commands related
to sink control will be rejected. Normally this should be an error,
but we're going to temporarily ignore this failure to allow existing
devices to update to FW with autofet disabled without fear of getting
stuck in a dead battery state.